main app completed

+++ src/app/api/webhooks/stripe/route.ts
@@ -1,4 +1,4 @@
-import { deleteAllFontsByClerkId, updateUserSubscription } from "@/db/crud";
+import { deleteAllFontsByClerkId, getUserIdByClerkId, updateUserSubscription } from "@/db/crud";
 import { priceIdToSubscriptionType } from "@/lib/stripe-helpers";
 import { sendToMixpanelServer } from "@/mixpanel/server-side";
 import { clerkClient } from "@clerk/nextjs/server";
@@ -13,8 +13,6 @@ export async function POST(req: NextRequest) {
   const payload = await req.text();
   const sig = req.headers.get("Stripe-Signature")!;
 
-  console.log(sig);
-  
   try {
     //=========={ Establish Event }==========//
     const event = stripe.webhooks.constructEvent(payload, sig, process.env.STRIPE_WEBHOOK_SECRET!); // eslint-disable-line
@@ -23,23 +21,26 @@ export async function POST(req: NextRequest) {
     const subscriptionType = priceIdToSubscriptionType(session.items.data[0].price.id);
 
     //=========={ DB syncing }==========//
-    const clerk = await clerkClient();
-    switch (event.type) {
-      case "customer.subscription.created":
-      case "customer.subscription.updated":
-        sendToMixpanelServer(userId, "subscription-updated", { subscriptionType, session_id: session.id });
-        await updateUserSubscription(userId, subscriptionType, session.id);
-        await clerk.users.updateUserMetadata(userId, { publicMetadata: { subscriptionType } }); // Update the subscription type in Clerk
-        break;
-      case "customer.subscription.deleted":
-        sendToMixpanelServer(userId, "subscription-deleted", { subscriptionType, session_id: session.id });
-        await updateUserSubscription(userId, "free", "");
-        await deleteAllFontsByClerkId(userId);
-        await clerk.users.updateUserMetadata(userId, { publicMetadata: { subscriptionType: "free" } });
-        break;
-      default:
-        sendToMixpanelServer(userId, "unhandled-event", { event: event.type });
-        break;
+    const userExistsQuery = await getUserIdByClerkId(userId);
+    if (userExistsQuery.length > 0) {
+      const clerk = await clerkClient();
+      switch (event.type) {
+        case "customer.subscription.created":
+        case "customer.subscription.updated":
+          sendToMixpanelServer(userId, "subscription-updated", { subscriptionType, session_id: session.id });
+          await updateUserSubscription(userId, subscriptionType, session.id);
+          await clerk.users.updateUserMetadata(userId, { publicMetadata: { subscriptionType } }); // Update the subscription type in Clerk
+          break;
+        case "customer.subscription.deleted":
+          sendToMixpanelServer(userId, "subscription-deleted", { subscriptionType, session_id: session.id });
+          await updateUserSubscription(userId, "free", "");
+          await deleteAllFontsByClerkId(userId);
+          await clerk.users.updateUserMetadata(userId, { publicMetadata: { subscriptionType: "free" } });
+          break;
+        default:
+          sendToMixpanelServer(userId, "unhandled-event", { event: event.type });
+          break;
+      }
     }
 
     return NextResponse.json({ status: "success", event: event.type }, { status: 200 });
